#11512: Rebasin refactoring branch
