+++ tests/sweep_framework/sweeps/eltwise/ternary/lerp/lerp.py
