+++ tests/sweep_framework/sweeps/eltwise/unary_backward/sin_bw/sin_bw.py
@@ -9,7 +9,7 @@ import torch
 import random
 import ttnn
 
-from tests.sweep_framework.utils import gen_shapes, sanitize_shape_rm
+from tests.sweep_framework.sweep_utils.utils import gen_shapes, sanitize_shape_rm
 from tests.tt_eager.python_api_testing.sweep_tests.generation_funcs import gen_func_with_cast_tt
 
 from tests.ttnn.utils_for_testing import check_with_pcc, start_measuring_time, stop_measuring_time
